Rename all statistics components to kebab-case convention
Rename all component files in statistics directory to follow
kebab-case naming convention for consistency with project standards.

Changes:
- Rename CategoryBarChart.tsx → category-bar-chart.tsx
- Rename ChartLegend.tsx → chart-legend.tsx
- Rename StatsSummaryCard.tsx → stats-summary-card.tsx
- Rename ParticipantBalanceRow.tsx → participant-balance-row.tsx
- Update all export paths in index.ts to use kebab-case imports

All type checks and linting pass successfully.
