+++ src/ui/components/statistics/category-bar-chart.tsx
