+++ src/ui/components/statistics/chart-legend.tsx
