+++ src/ui/components/statistics/participant-balance-row.tsx
